requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
httpx==0.26.0
langchain==0.1.0
langchain-community==0.0.13
chromadb==0.4.22
groq==0.4.1
google-generativeai==0.3.2
PyMuPDF==1.23.8
python-docx==1.1.0
tiktoken==0.5.2
